requirements.txt
pymupdf
rapidocr-onnxruntime
ebooklib
beautifulsoup4
lxml
pillow
